bump dependencies, rearrange logic
