+++ src/main/kotlin/com/kortov/bootigram/bots/handlers/QuizHandler.kt
@@ -3,4 +3,5 @@ package com.kortov.bootigram.bots.handlers
 import org.springframework.stereotype.Component
 
 @Component
-class QuizHandler
+class QuizHandler(val fileHandler: FileHandler) {
+}
